refactor(settings): move mr-rework accessors, const, validator to settings_mr_rework.go (PRD #1021 M3)
